fix: fix handcoded q13 for read_relation (haog/gqe-python!126)
this is related to this [MR!117](https://gitlab-master.nvidia.com/haog/gqe-python/-/merge_requests/117). I forgot to update Q13 since at that time Q13 was newly added and I didn't sync with the upstream for the latest.

After this fix, Q13 handcoded would be able to run with parameter sweep and also print the physical plans using the gqe-tpch-pretuned command.

Thanks @siyuanl for pointing this out.