Fix the program return code for table registration failures. (haog/gqe-python!139)
Currently, when a parquet file cannot be read, for example, if the file is missing, those tests will be skipped.  This leads to no errors and the test passes.

This is bad because it makes it appear as if CI is passing when really we didn't test anything.

We already have detection of missing parquet input files.  When those files are missing, we should count that as an error.  Having an invalid dataset should be a reported error.