Add function to track unrecoverable CUDA errors to allow process to quit (haog/gqe-python!170)
The CUDA runtime system treats asynchronous errors, such as `cudaIllegalMemoryAccess`, as unrecoverable. The CUDA context becomes corrupted and requires ending the process.  cuDF wraps such errors as exceptions, for which `gqe-python` attempts to continue processing.

This MR adds a function to check if an error is unrecoverable, and if it is, allows us to end the process, as no further query can be computed successfully. For sandboxing, this enables the sandbox process to be restarted, continuing execution as intended.