Add CI tests for SQL views in the experiment database (haog/gqe-python!213)
The SQL Views in the SQLite schema break now and again, when MRs parameters add or remove parameters but don't update the View queries.

This MR adds `pytest` tests that take the output of the TPC-H parameter sweep, and check that the Views work as intended.

In future, the tests could be expanded to check that, when configured, each parameter correctly shows up in the SQLite database. See https://gitlab-master.nvidia.com/haog/gqe-python/-/issues/29.